requirements.txt
aiogram==3.*
aiohttp
python-dotenv
googletrans
asyncio
aioschedule